JBRULES-2835: revert ksession id from long to int

+++ drools-api/src/main/java/org/drools/persistence/jpa/KnowledgeStoreService.java
@@ -27,7 +27,7 @@ public interface KnowledgeStoreService {
                                                          KnowledgeSessionConfiguration configuration,
                                                          Environment environment);
 
-    StatefulKnowledgeSession loadStatefulKnowledgeSession(long id,
+    StatefulKnowledgeSession loadStatefulKnowledgeSession(int id,
                                                           KnowledgeBase kbase,
                                                           KnowledgeSessionConfiguration configuration,
                                                           Environment environment);
